+++ src/com/donny/dendrofinance/gui/menu/reports/BalanceSheetGui.java
@@ -37,7 +37,7 @@ public class BalanceSheetGui extends RegisterFrame {
             ENTER.addActionListener(event -> updateTable());
 
             PANE = DendroFactory.getTable(new String[]{
-                    "Account", "Debit", "Credit", "Tracking", "Ghost"
+                    "Account", "Debit", "Credit", "Ghost", "Tracking"
             }, new Object[][]{}, false);
             TABLE = (JTable) PANE.getViewport().getView();
             TABLE_ACCESS = (DefaultTableModel) TABLE.getModel();
@@ -120,7 +120,6 @@ public class BalanceSheetGui extends RegisterFrame {
                 if (acc.get(a).compareTo(compare) >= 0 || acc.get(a).compareTo(compare.multiply(BigDecimal.valueOf(-1))) <= 0) {
                     typ.add(a.getBroadAccountType(), acc.get(a));
                     accTyp.add(a.getAccountType(), acc.get(a));
-                    String search = SEARCH.getText();
                     if (
                             a.getName().toLowerCase().contains(SEARCH.getText().toLowerCase()) ||
                                     a.getCurrency().toString().toLowerCase().contains(SEARCH.getText().toLowerCase()) ||
@@ -134,10 +133,10 @@ public class BalanceSheetGui extends RegisterFrame {
                             case LIABILITY, EQUITY_PLUS, REVENUE -> TABLE_ACCESS.addRow(new String[]{
                                     a.getName(), "", a.getCurrency().encode(acc.get(a)), "", ""
                             });
-                            case TRACKING -> TABLE_ACCESS.addRow(new String[]{
+                            case GHOST -> TABLE_ACCESS.addRow(new String[]{
                                     a.getName(), "", "", a.getCurrency().encode(acc.get(a)), ""
                             });
-                            case GHOST -> TABLE_ACCESS.addRow(new String[]{
+                            case TRACKING -> TABLE_ACCESS.addRow(new String[]{
                                     a.getName(), "", "", "", a.getCurrency().encode(acc.get(a))
                             });
                         }